backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ICheered <tjeerd992@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.67.0"
uvicorn = "^0.14.0"
pydantic = "^1.8.2"
tinydb = "^4.5.1"
email-validator = "^1.1.3"
passlib = "^1.7.4"
python-jose = "^3.3.0"
argon2-cffi = "^20.1.0"
python-multipart = "^0.0.5"
fastapi-mail = "^0.4.0"
pymongo = "^3.12.0"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = {version = "^21.7b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
